refactor(security): segregate IBudgetLineSecurity to break dependency cycle
HasBudgetLinePermission never used IAuthorizationService; it lived in
SecurityRepository only by placement, dragging that class's dependency
on IAuthorizationService into the handler's construction graph. The
check moves unchanged to BudgetLineSecurity behind IBudgetLineSecurity,
which the handler constructor-injects. ISecurity keeps its family and
admin assertions exactly as before - no security logic is duplicated.
